ArticleDetailedViewController - add a timeout:
If there is a network (even "100 % Loss" profile) - web view is loading forever,
not calling any of delegates methods (webView:didFinishLoad: or webViewDidFailWithError: )),
so GUI is disabled and activity indicator is animating forever.
Add a timer (1 m.) to check, if we're still loading and cancel all operations after timeout.
Also, hide error HUD when switching back from the web-view to the readability view (otherwise it's always on top
of everything).