Keep it simple, .... smart.
Instead of fighting for memory in every possible view and controller with
f..g twisted logic and additional super-complications (and more and more and
more testing required) ... just replace the top view controller in a slider
with a feed view controller thus deleting and reloading everything.
Well, it can happen that I'll receive this message again, while processing the
previous one ... but in this case nothing can help.