Use a shorter timeout when updating the menu from the remote plist:
Since we check for the notifications only after menu was reloaded,
there is not reason to wait the standard 60 seconds for the menu update to fail (if there are network-related probles) -
if the app started from the notificatino center, it's good to see what was new as soon as possible (well, 10 s. is enough).