Change the APN machinery. Since we can "accumulate" notifications,
I need the number of new items from the payload. Also, hint
is not removed now when menu item selected: since we are caching feeds,
it's possible to open cached feed without actual refresh and so
I can not immediately hide a hint view on this menu item.
Unfortunately, I also have to "propagate" APN handling to content providers,
since it's a content provider who deals with controller and controller
also have to know about the number of new items to show a hint in a
navigation bar.

TODO: since at the beginning we did not have APNs and all three parts
(content providers, menu items, menu item views) were quite decoupled,
things are getting quite ugly/repetitive and twisted - to be redesigned in
the future version.