Change the APN hint's machinery:
Do not accumulate hints: it possible (due to bugs in iOS?) that notification is not removed
from the notification center (actually, on iOS 6 there is no even API for this, just resetting
the number of badges attached to the app's icon - really nice "API" :) - in this case it possible
to "accumulate" the same notification infinite number of times. Instead - reset number shown in an APN hint view
(and accumulate correctly only at the level of group menu item).