Synchronous connection in a background thread
prevents this thread's cancellation - operation still sits in a queue
until download has finished/failed in this can take quite a lot of time.
While controller is dead for a long time now, operation (and thus additional thread)
is still active and taking resources. Make it more flexible - make a controller class
to download xml and later invoke a thread for a parsing operation so that connection
can be closed/cancelled earlier.