Evil user can always break an app if he really wants so :(((
In photo collections it's possible to tap "Back to collections" (thus closing the selected collection) and
almost immediately on some image in the collection with completely unpredictable results - stack/unstack animation is executed
on a main GUI thread, but not immediately after it was started and somehow the second tap can be processed somewhere in the middle.

This is disgasting.